Initial commit: XG Proyect v3.x.x fork with enhancements
- PHP 8.2+ migration (Docker, CI, composer)
- Code quality: PHPStan L2 clean, PHP CS Fixer, PHPUnit tests
- Type hinting migration (PHPDoc -> native PHP 8 types)
- Attack mission bug fixes (battle engine, report rendering)
- Bot system: LLM integration, heat system, alliance/attack/diplomacy modules
- Detailed roadmap for bot improvements (planes/bot_roadmap_2026_05.md)

+++ app/Services/Formulas/Expedition.php
@@ -47,7 +47,7 @@ class Expedition
 
     public function calculateExpeditionPoints(int $structuralIntegrity): int
     {
-        return ($structuralIntegrity * 5 / 1000);
+        return (int) ($structuralIntegrity * 5 / 1000);
     }
 
     public function getExpeditionResult(): string
